+++ config.go
@@ -3,6 +3,7 @@ package neat
 import (
 	"database/sql"
 	"fmt"
+	"io/fs"
 	"strings"
 	"time"
 
@@ -58,7 +59,8 @@ type ConnectionConfig struct {
 	NameReplacer any
 	Read         []ReplicaConfig
 	Write        []ReplicaConfig
-	Tables       any // GODB: godb.Tables or []godb.Table; ignored by other drivers
+	Tables       any   // GODB: godb.Tables or []godb.Table; ignored by other drivers
+	FS           fs.FS // CSVDB/JSONDB/XMLDB: embedded filesystem (embed.FS / fs.FS); ignored by other drivers
 }
 
 // String returns a string representation of ConnectionConfig with password masked.
@@ -380,6 +382,7 @@ func New(cfg DBConfig, opts ...database.Option) (*database.Database, error) {
 			NoLowerCase:  conn.NoLowerCase,
 			NameReplacer: conn.NameReplacer,
 			Tables:       conn.Tables,
+			FS:           conn.FS,
 		}
 		for _, r := range conn.Read {
 			dbConn.Read = append(dbConn.Read, db.ReplicaConfig{